extracted relevant logic to a MapConverter class
